requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
PyMySQL==1.1.0
mysqlclient==2.2.0
Werkzeug==2.3.6
marshmallow==3.19.0
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
chardet==5.1.0
faker==19.3.0
requests==2.31.0 